Hooked up crud and display validation results

+++ ShopCheckDb/ShopCheckService.cs
@@ -21,7 +21,7 @@ namespace ShopCheckDb
         {
             DbSet<Product> allProducts = _db.Products;
             
-            return allProducts.ToArray();
+            return allProducts.ToList();
 
         }
 
@@ -39,39 +39,82 @@ namespace ShopCheckDb
 
         public ServiceResult DeleteProduct(int id)
         {
-            Product productToDelete = ReadProduct(id);
-            _db.Products.Remove(productToDelete);
-            _db.SaveChanges();
-            return null; 
+            Product? productToDelete = ReadProduct(id);
+            ServiceResult sr = new ServiceResult();
+
+            if(productToDelete == null)
+            {
+                sr.ValidationErrors.Add("DELETE", "Cannot find " + id);
+            }
+
+            if(sr.IsValid)
+            {
+                _db.Products.Remove(productToDelete);
+                _db.SaveChanges();
+            }
+            return sr; 
 
         }
 
         public ServiceResult UpdateProduct(Product product)
         {
-            Product existing = ReadProduct(product.Id);
-            existing.Name = product.Name;
-            existing.Url = product.Url;
-            existing.MinStock = product.MinStock;
-            _db.SaveChanges();
-            return null;
+            var sr = Validate(product);
+
+
+
+            if (sr.IsValid)
+            {
+                Product? existing = ReadProduct(product.Id);
+
+                if(existing == null)
+                {
+                    sr.ValidationErrors.Add("NULL", $"Cannot read product {product.Id}");
+                    return sr;
+                }
+
+                existing.Name = product.Name;
+                existing.Url = product.Url;
+                existing.MinStock = product.MinStock;
+                _db.SaveChanges();
+            }
+            return sr;
+        }
+
+        private ServiceResult Validate(Product product)
+        {
+            ValidationContext validationContext = new ValidationContext(product);
+            List<ValidationResult> results = new List<ValidationResult>();
+            ServiceResult sr = new ServiceResult();
+            if (!Validator.TryValidateObject(product, validationContext, results, true))
+            {
+                int i = 0;
+                foreach (var r in results)
+                {
+                    sr.ValidationErrors.Add(i.ToString(), r.ErrorMessage);
+                    i++;
+                }
+                return sr;
+            }
+
+            return sr;
+
         }
 
         public ServiceResult CreateProduct(Product product)
         {
-            
 
-            ValidationContext validationContext = new ValidationContext(product);
-            List<ValidationResult> results = new List<ValidationResult>();
 
-            if(!Validator.TryValidateObject(product, validationContext, results, true))
+
+            var sr = Validate(product);
+
+
+            if (sr.IsValid)
             {
-                return null;
+                product.Id = 0;
+                _db.Products.Add(product);
+                _db.SaveChanges();
             }
-
-            product.Id = 0;
-            _db.Products.Add(product);
-            _db.SaveChanges();
-            return null;
+            return sr;
         }
